Fix FFP/vbcc compatibility for float handling in runtime library
Change function signatures and variable types from float to long in
FFP-related code to treat FFP values as opaque 32-bit quantities.
This prevents vbcc from misinterpreting FFP format as IEEE format.

Changes:
- val.c: Change val() return type and local vars to long, replace
  float literals with SPFlt() calls, add SPNeg() for negative values
- pow.c: Change power() signature and extern declarations to long
- print.c: Change Usingleprint() parameter to long
- ffpstr.c: Change strsingle() parameter and afp() return to long

Includes rebuilt db.lib and startup.lib, plus FFP verification test.

+++ src/lib/c/ffpstr.c
@@ -34,11 +34,11 @@ extern	long	decimal_places;	/* number of places to round to -- default=8 */
 
 /* FFP math functions */
 long 	SPTst();
-float	afp();
+long	afp();    /* returns FFP as raw 32-bit */
 
 /* strsingle() */
 char 	*strsingle(fnum)
-float 	fnum;
+long 	fnum;    /* FFP value as raw 32-bit */
 {
 static 	char 	fpabuf[15];
 static	char	ex_buf[3];
